refactor: консолидация констант DISP_WIDTH и DISP_HEIGHT
- Удалено дублирование констант между io.rs и game/constants.rs
- Константы определены в game/constants.rs как вычисляемые
- Добавлен переэкспорт из io.rs для обратной совместимости
- Исправлены типы с u16 на usize для совместимости
- Обновлены тесты test_io.rs и test_cast_safety.rs

+++ src/app/application.rs
@@ -93,7 +93,7 @@ impl Application {
             e
         })?;
 
-        if width < DISP_WIDTH || height < DISP_HEIGHT {
+        if (width as usize) < DISP_WIDTH || (height as usize) < DISP_HEIGHT {
             let err = std::io::Error::new(
                 std::io::ErrorKind::InvalidInput,
                 format!(
